Separate macro to its own crate
