+++ src/lib.rs
@@ -1,17 +1,3 @@
-extern crate proc_macro; // Needed even though it's the 2018 edition. See https://github.com/idanarye/rust-typed-builder/issues/57.
-
-use proc_macro2::TokenStream;
-
-use syn::parse::Error;
-use syn::spanned::Spanned;
-use syn::{parse_macro_input, DeriveInput};
-
-use quote::quote;
-
-mod field_info;
-mod struct_info;
-mod util;
-
 /// `TypedBuilder` is not a real type - deriving it will generate a `::builder()` method on your
 /// struct that will return a compile-time checked builder. Set the fields using setters with the
 /// same name as the struct's fields and call `.build()` when you are done to create your object.
@@ -162,48 +148,7 @@ mod util;
 ///     `param1: Type1, param2: Type2 ...` instead of the field type itself. The parameters are
 ///     transformed into the field type using the expression `expr`. The transformation is performed
 ///     when the setter is called.
-#[proc_macro_derive(TypedBuilder, attributes(builder))]
-pub fn derive_typed_builder(input: proc_macro::TokenStream) -> proc_macro::TokenStream {
-    let input = parse_macro_input!(input as DeriveInput);
-    match impl_my_derive(&input) {
-        Ok(output) => output.into(),
-        Err(error) => error.to_compile_error().into(),
-    }
-}
-
-fn impl_my_derive(ast: &syn::DeriveInput) -> Result<TokenStream, Error> {
-    let data = match &ast.data {
-        syn::Data::Struct(data) => match &data.fields {
-            syn::Fields::Named(fields) => {
-                let struct_info = struct_info::StructInfo::new(ast, fields.named.iter())?;
-                let builder_creation = struct_info.builder_creation_impl()?;
-                let conversion_helper = struct_info.conversion_helper_impl();
-                let fields = struct_info
-                    .included_fields()
-                    .map(|f| struct_info.field_impl(f))
-                    .collect::<Result<TokenStream, _>>()?;
-                let required_fields = struct_info
-                    .included_fields()
-                    .filter(|f| f.builder_attr.default.is_none())
-                    .map(|f| struct_info.required_field_impl(f));
-                let build_method = struct_info.build_method_impl();
-
-                quote! {
-                    #builder_creation
-                    #conversion_helper
-                    #fields
-                    #(#required_fields)*
-                    #build_method
-                }
-            }
-            syn::Fields::Unnamed(_) => return Err(Error::new(ast.span(), "TypedBuilder is not supported for tuple structs")),
-            syn::Fields::Unit => return Err(Error::new(ast.span(), "TypedBuilder is not supported for unit structs")),
-        },
-        syn::Data::Enum(_) => return Err(Error::new(ast.span(), "TypedBuilder is not supported for enums")),
-        syn::Data::Union(_) => return Err(Error::new(ast.span(), "TypedBuilder is not supported for unions")),
-    };
-    Ok(data)
-}
+pub use typed_builder_macro::TypedBuilder;
 
 // It'd be nice for the compilation tests to live in tests/ with the rest, but short of pulling in
 // some other test runner for that purpose (e.g. compiletest_rs), rustdoc compile_fail in this
