+++ typed-builder-macro/Cargo.toml
@@ -1,0 +1,12 @@
+[package]
+name = "typed-builder-macro"
+version = "0.14.0"
+edition = "2018"
+
+[lib]
+proc-macro = true
+
+[dependencies]
+syn = { version = "2", features = ["full", "extra-traits"] }
+quote = "1"
+proc-macro2 = "1"
