+++ typed-builder-macro/src/struct_info.rs
